+++ test/select.test.ts
@@ -1,5 +1,5 @@
 import { test, expect } from '@jest/globals';
-import Select from '../src/components/Select.js';
+import Select from '../src/Select.js';
 
 test('button', () => {
   expect(Select({ title: 'A title', options: [], defaultValue: '', onChange: () => { console.log('test') } })).toBeDefined()
