+++ test/separator.test.ts
@@ -1,5 +1,5 @@
 import { test, expect } from '@jest/globals';
-import Separator from '../src/components/Separator.js'
+import Separator from '../src/Separator.js'
 
 test('button', () => {
   expect(Separator({ title: 'A title', onClick: () => { console.log('a test') } })).toBeDefined()
